+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 React Intersection adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0] 2019-12-05
+
+- Updated dependencies
+- Added use of HTMLImageElement.decode()
+- Abort image download on unmount
+
 ## [1.0.1] 2018-02-20
 
 ### Fixes